com.unity.serialization@3.1.1 ## [3.1.1] - 2023-05-10 ### Changed * Updated `com.unity.collections` to version `2.1.4`.
## [3.1.0] - 2023-04-26
### Added
* Added `JsonSerializationParameters.StringEscapeHandling` to optionally disable escape character handling. The default value is true.

### Changed
* Updated `com.unity.collections` to version `2.1.2`.
* Updated `JsonSerialization` to allow primitive and string types to be adapted using `IJsonAdapter{T}`.

### Removed
* Removed dependency on `com.unity.test-framework.performance`.

+++ Runtime/Unity.Serialization/Json/Adapters/Internal/JsonAdapter+Primitives.cs
@@ -28,7 +28,7 @@ namespace Unity.Serialization.Json
         void IJsonAdapter<float>.Serialize(in JsonSerializationContext<float> context, float value) => context.Writer.WriteValue(value);
         void IJsonAdapter<double>.Serialize(in JsonSerializationContext<double> context, double value) => context.Writer.WriteValue(value);
         void IJsonAdapter<bool>.Serialize(in JsonSerializationContext<bool> context, bool value) => context.Writer.WriteValueLiteral(value ? "true" : "false");
-        void IJsonAdapter<char>.Serialize(in JsonSerializationContext<char> context, char value) => context.Writer.WriteValue(value);
+        void IJsonAdapter<char>.Serialize(in JsonSerializationContext<char> context, char value) => context.Writer.WriteValue((int) value);
         void IJsonAdapter<string>.Serialize(in JsonSerializationContext<string> context, string value) => context.Writer.WriteValue(value);
         
         sbyte IJsonAdapter<sbyte>.Deserialize(in JsonDeserializationContext<sbyte> context)
